--- a/build/configuration/pom.xml
+++ b/build/configuration/pom.xml
@@ -141,7 +141,7 @@
       <version.infinispan>15.0.0-SNAPSHOT</version.infinispan>
       <version.infinispan.arquillian>1.2.0.Beta3</version.infinispan.arquillian>
       <version.infinispan.doclets>1.4.0.Final</version.infinispan.doclets>
-      <version.infinispan.maven-plugins>1.0.3.Final</version.infinispan.maven-plugins>
+      <version.infinispan.maven-plugins>1.0.4.Final</version.infinispan.maven-plugins>
       <version.io.agroal>2.1</version.io.agroal>
       <version.io.lettuce>6.2.0.RELEASE</version.io.lettuce>
       <version.io.mashona>1.0.0</version.io.mashona>