https://issues.redhat.com/browse/ISPN-14809

Basic changes boil down to publishing a value to the out List so the ByteToMessageDecoder knows that if auto read is disabled it can actually disable it. This allows us to remove the workaround code that was done before where it would replay the internal buffer when auto read was reenabled on our side. Unfortunately, this means the internal buffer could grow unbounded if a client kept writing to the socket.