@wburns 2023-05-11
The decoder is very slightly slower than before, but that is to expected as its hot path got a little longer having to be passed down the netty pipeline. There isn't much here can do to change that. The auto read path however is much simpler than before, not to mention that we don't have to replay the BytetoMessage internal buffer anymore.

MAIN
```
Benchmark                        (decoder)  (messageCount)  Mode  Cnt  Score   Error  Units
MyBenchmark.testGetPerf            CURRENT             N/A  avgt    5  0.756 ± 0.026  us/op
MyBenchmark.testGetPipelinePerf    CURRENT               2  avgt    5  1.155 ± 0.040  us/op
MyBenchmark.testGetPipelinePerf    CURRENT              25  avgt    5  6.474 ± 0.038  us/op
MyBenchmark.testSetPerf            CURRENT             N/A  avgt    5  0.797 ± 0.025  us/op
MyBenchmark.testSetPipelinePerf    CURRENT               2  avgt    5  1.252 ± 0.023  us/op
MyBenchmark.testSetPipelinePerf    CURRENT              25  avgt    5  7.006 ± 0.127  us/op
```
NEW
```
Benchmark                        (decoder)  (messageCount)  Mode  Cnt  Score   Error  Units
MyBenchmark.testGetPerf            CURRENT             N/A  avgt    5  0.803 ± 0.012  us/op
MyBenchmark.testGetPipelinePerf    CURRENT               2  avgt    5  1.294 ± 0.016  us/op
MyBenchmark.testGetPipelinePerf    CURRENT              25  avgt    5  7.342 ± 0.101  us/op
MyBenchmark.testSetPerf            CURRENT             N/A  avgt    5  0.819 ± 0.005  us/op
MyBenchmark.testSetPipelinePerf    CURRENT               2  avgt    5  1.339 ± 0.023  us/op
MyBenchmark.testSetPipelinePerf    CURRENT              25  avgt    5  7.816 ± 0.026  us/op
```