@wburns 2023-05-16
Updated to use 1.0.4.Final for the protocol generator.